Match Java type 'long' to result of postgres count()
Postgres docs say that count() yields Postgres numerical type bigint [1]
which is 8 bytes according to [2].

This means that our Dao classes and pagination support should upgrade from
using 'int' to 'long' where applicable.

[1] https://www.postgresql.org/docs/8.2/functions-aggregate.html
[2] https://www.postgresql.org/docs/9.1/datatype-numeric.html

+++ textrepo-app/src/main/java/nl/knaw/huc/core/Page.java
@@ -7,10 +7,10 @@ import java.util.List;
 public class Page<T> {
 
   private final List<T> items;
-  private final int total;
+  private final long total;
   private final PageParams params;
 
-  public Page(List<T> items, int total, PageParams params) {
+  public Page(List<T> items, long total, PageParams params) {
     this.items = items;
     this.total = total;
     this.params = params;
@@ -20,7 +20,7 @@ public class Page<T> {
     return items;
   }
 
-  public int getTotal() {
+  public long getTotal() {
     return total;
   }
